chore: update

+++ apps/bodhi-service/src/modules/files/files.controller.ts
@@ -1,4 +1,5 @@
-import { Body, Delete, Param, Post, Req, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Delete, NotFoundException, Param, Post, Req } from '@nestjs/common';
+import { UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
 import { Controller, Get, HttpException, HttpStatus } from '@nestjs/common';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { ApiBearerAuth, ApiSecurity, ApiTags } from '@nestjs/swagger';
@@ -51,6 +52,7 @@ export class FilesController {
   async upload(@Req() req: RequestWithUser, @UploadedFiles() files, @Body() body: UploadFileReq) {
     const { user_id, client_user_id = '' } = req.user; // from jwt or apikey
     const { purpose } = body;
+    const expires_at = new Date(Date.now() + 1000 * 60 * 60 * 24 * 30); // 30 days
 
     try {
       // 计算并检查hash
@@ -63,7 +65,8 @@ export class FilesController {
           const size = file.size;
           const name = Buffer.from(file.originalname, 'latin1').toString('utf8');
 
-          return this.service.uploadFile(file, { hash, name, mimetype, size, user_id, client_user_id }, purpose);
+          const opts = { hash, name, mimetype, size, expires_at, user_id, client_user_id };
+          return this.service.uploadFile(file, opts, purpose);
         }),
       );
     } catch (err) {
@@ -74,15 +77,15 @@ export class FilesController {
   @Get(':id')
   @ApiOperation({ summary: 'Get file detail', description: 'Get file detail' })
   @ApiResponse({ status: 200, description: 'success', type: FileDto })
-  async get(@Req() req: RequestWithUser, @Param('id') id: number): Promise<FileDto> {
+  async get(@Req() req: RequestWithUser, @Param('id') file_id: string): Promise<FileDto> {
     const { user_id, client_user_id = '' } = req.user; // from jwt or apikey
 
     try {
-      const file = await this.service.findActiveFilesById(id, user_id, client_user_id);
+      const id = this.file.decodeId(file_id);
+      const file = await this.service.findActiveById(id, user_id, client_user_id);
       const url = `https://s.alidraft.com${file.path}`;
       delete file.path;
 
-      const file_id = this.file.encodeId(file.id);
       return { ...file, id: file_id, url };
     } catch (err) {
       throw new HttpException(err.message, HttpStatus.FORBIDDEN);
@@ -92,11 +95,16 @@ export class FilesController {
   @Delete(':id')
   @ApiOperation({ summary: 'Delete File', description: 'Delete File' })
   @ApiResponse({ status: 200, description: 'success' })
-  async delete(@Req() req: RequestWithUser, @Param('id') id: number) {
+  async delete(@Req() req: RequestWithUser, @Param('id') file_id: string) {
     const { user_id, client_user_id = '' } = req.user; // from jwt or apikey
 
     try {
-      await this.service.delete(id, user_id, client_user_id);
+      const id = this.file.decodeId(file_id);
+      const file = await this.service.findActiveById(id, user_id, client_user_id);
+      if (file && ['active', 'expired', 'created'].includes(file.state)) {
+        return this.service.delete(id, user_id, client_user_id);
+      }
+      throw new NotFoundException();
     } catch (err) {
       throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
     }
